requirements.txt
flask==3.0.3
flask-cors==4.0.1
mongoengine==0.29.1
marshmallow==3.21.1
python-dotenv==1.0.1
PyJWT==2.8.0
gunicorn==23.0.0
cloudinary==1.41.0
pillow==10.4.0
